feat: Add support for intentionally invalid/failing test vectors (#1)
Also separates out the concept of "decryption generation" vectors from encryption vectors: the former is a specification for how to generate decryption tests, whereas the second is a specification of encryption tests. They overlap substantially but are not the same!
