+++ ci/prettify.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+case "${1}" in
+  write)
+    npx prettier --config .prettierrc.toml --write -- '**/*.md'
+    ;;
+  check)
+    npx prettier --config .prettierrc.toml --check -- '**/*.md'
+    ;;
+  *)
+    echo "mode required!"
+    echo "${0} [write/check]"
+    exit 1
+    ;;
+esac
